// src/generated/models/PH_UsersModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */

export interface EditorValue {
  "@odata.type": string;
  Claims: string;
  DisplayName: string;
  Email: string;
  Picture: string;
  Department: string;
  JobTitle: string;
}

export interface AuthorValue {
  "@odata.type": string;
  Claims: string;
  DisplayName: string;
  Email: string;
  Picture: string;
  Department: string;
  JobTitle: string;
}

export interface PH_UsersBase {
  ID?: number;
  Title?: string;
  DisplayName?: string;
  Email?: string;
  Jobtitle?: string;
  Department?: string;
  Location?: string;
  ManagerPath?: string;
  Manager?: string;
  Review1Complete?: boolean;
  Review2Complete?: boolean;
  AccountEnabled?: boolean;
  Modified?: string;
  ManualActive?: boolean;
  Photo?: string;
  "Editor#Claims"?: string;
  CultureRank?: number;
  KPIRank?: number;
  OverallRank?: number;
  KPIAvg?: number;
  CultureAvg?: number;
  OverallAvg?: number;
  OData__ColorTag?: string;
  ComplianceAssetId?: string;
  Created?: string;
  "Author#Claims"?: string;
  "{Identifier}"?: string;
  "{IsFolder}"?: boolean;
  "{Thumbnail}"?: {
    "@odata.type"?: string;
    // URL to the small, highly compressed size thumbnail of the item, if available.
    Small?: string;
    // URL to the standard size thumbnail of the item, if available.
    Medium?: string;
    // URL to the largest size thumbnail of the item, if available.
    Large?: string;
  };
  "{Link}"?: string;
  "{Name}"?: string;
  "{FilenameWithExtension}"?: string;
  "{Path}"?: string;
  "{FullPath}"?: string;
  "{ModerationStatus}"?: string;
  "{ModerationComment}"?: string;
  "{ContentType}#Id"?: string;
  "{ContentType}"?: {
    "@odata.type"?: string;
    Id?: string;
    Name?: string;
  };
  "{HasAttachments}"?: boolean;
  "{Attachments}@odata.type"?: string;
  "{Attachments}"?: {
   "@odata.type"?: string;
   Id?: string;
   AbsoluteUri?: string;
   DisplayName?: string;
 }[];
  "{VersionNumber}"?: string;
  "{TriggerWindowStartToken}"?: string;
  "{TriggerWindowEndToken}"?: string;
}

export interface PH_UsersRead extends PH_UsersBase {
  Editor?: EditorValue;
  Author?: AuthorValue;
}

export interface PH_UsersWrite extends PH_UsersBase {
  Editor?: EditorValue;
  Author?: AuthorValue;
}